feat: [SCRUM-97] refactor code
