+++ src/app/routes/PrivateRoute.jsx
@@ -1,6 +1,5 @@
 import HomePage from '@pages/HomePage.jsx'
 import GrammarVocabPage from '@pages/GrammarVocabPage.jsx'
-import InstructionsGrammarPage from '@pages/InstructionsGrammarPage.jsx'
 import WritingTestPage from '@pages/WritingTestPage.jsx'
 import IntroductionScreen from '@pages/IntroductionScreen'
 import { ProtectedRoute } from './ProtectedRoute/ProtectedRoute.jsx'
@@ -20,12 +19,8 @@ const PrivateRoute = [
         element: <HomePage />
       },
       {
-        path: 'grammar-vocab',
+        path: 'grammarvocab',
         element: <GrammarVocabPage />
-      },
-      {
-        path: 'grammar-instructions',
-        element: <InstructionsGrammarPage />
       },
       {
         path: 'listening',
